requirements.txt
crewai>=0.11.0
openai>=1.0.0
streamlit>=1.32.0
python-dotenv>=1.0.0
duckduckgo-search>=4.1.1
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
pydantic>=2.0.0
requests>=2.31.0
markdown>=3.5.0 